requirements.txt
opencv-python-headless>=4.5.5.64
mtcnn==0.1.1
tensorflow-cpu==2.16.1
numpy>=1.26.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
python-jose==3.3.0
python-dotenv==1.0.0
jinja2==3.1.3 